refactor(scripts): #72 2.2-4 — extract EODHD planner+runner to scripts/validate/eodhd.py
Move EODHD constants (six EODHD_*_READS sets,
EODHD_SUBTYPE_READS / _PARSERS, EODHD_ENUM_FIELDS_BY_SUBTYPE), the
planner, _flatten_rows, diff_eodhd_row, try_parse_eodhd, and
run_eodhd_probe into scripts/validate/eodhd.py (372 LOC). The
validator drops 3,732 → 3,406 LOC.

Cross-cutting helper bump in `_shared.py`: `_render_params` (used by
both the EODHD runner and the validator's dry-run printer) is
promoted to a public `render_params`. The same pattern as the ISO
date helpers in 2.2-3 — keeps the new EODHD module from back-importing
to the parent script.

Tests: 9 scaffold tests load the validator unchanged at 135 passed;
broader `-k "scaffold or validate or scripts"` subset at 695 passed.
Manual `--provider eodhd` dry-run plans 9 probes correctly with the
new shared `render_params` formatting query strings.

Refs #72.
